+++ packages/synapse-interface/utils/types/index.tsx
@@ -71,6 +71,9 @@ type QuoteQuery = {
 
 export type BridgeQuote = {
   id: string
+  inputAmountForQuote: string
+  originTokenForQuote: Token
+  destTokenForQuote: Token
   outputAmount: bigint
   outputAmountString: string
   routerAddress: string
@@ -86,6 +89,7 @@ export type BridgeQuote = {
   timestamp: number
   originChainId: number
   destChainId: number
+  requestId: number
 }
 
 interface TokensByChain {