+++ services/rfq/relayer/relapi/client.go
@@ -15,12 +15,15 @@ import (
 // RelayerClient is the interface for the relayer client.
 type RelayerClient interface {
 	Health(ctx context.Context) (ok bool, err error)
+	// Deprecated: use GetQuoteRequestByTxHash
 	GetQuoteRequestStatusByTxHash(ctx context.Context, hash string) (*GetQuoteRequestStatusResponse, error)
+	// Deprecated: use GetQuoteRequestStatusByTxID
 	GetQuoteRequestStatusByTxID(ctx context.Context, hash string) (*GetQuoteRequestStatusResponse, error)
 	RetryTransaction(ctx context.Context, txhash string) (*GetTxRetryResponse, error)
 	Withdraw(ctx context.Context, req *WithdrawRequest) (*WithdrawResponse, error)
 	GetTxHashByNonce(ctx context.Context, req *GetTxByNonceRequest) (*TxHashByNonceResponse, error)
 	GetQuoteRequestByTXID(ctx context.Context, txid string) (*GetQuoteRequestResponse, error)
+	GetQuoteRequestByTxHash(ctx context.Context, txhash string) (*GetQuoteRequestResponse, error)
 }
 
 type relayerClient struct {
@@ -141,7 +144,6 @@ func (r *relayerClient) GetTxHashByNonce(ctx context.Context, req *GetTxByNonceR
 		SetQueryParam("chain_id", fmt.Sprintf("%d", req.ChainID)).
 		SetQueryParam("nonce", fmt.Sprintf("%d", req.Nonce)).
 		Get(getTxHashByNonceRoute)
-
 	if err != nil {
 		return nil, fmt.Errorf("failed to get tx hash by nonce: %w", err)
 	}
@@ -161,6 +163,23 @@ func (r *relayerClient) GetQuoteRequestByTXID(ctx context.Context, txid string) 
 		Get(getRequestByTxID)
 	if err != nil {
 		return nil, fmt.Errorf("failed to get quote request by tx id: %w", err)
+	}
+
+	if resp.StatusCode() != http.StatusOK {
+		return nil, fmt.Errorf("unexpected status code: %d", resp.StatusCode())
+	}
+
+	return &res, nil
+}
+
+func (r *relayerClient) GetQuoteRequestByTxHash(ctx context.Context, txHash string) (*GetQuoteRequestResponse, error) {
+	var res GetQuoteRequestResponse
+	resp, err := r.client.R().SetContext(ctx).
+		SetQueryParam("hash", txHash).
+		SetResult(&res).
+		Get(getRequestByTxHash)
+	if err != nil {
+		return nil, fmt.Errorf("failed to get quote request by tx hash: %w", err)
 	}
 
 	if resp.StatusCode() != http.StatusOK {